feat(acme): implement ACME DNS management and ECH support
- Add certificate management functionality in `certificate.go` for ACME.
- Introduce ECH (Encrypted Client Hello) configuration management in `ech.go`.
- Implement ENS (Ethereum Name Service) gasless hostname management in `ens.go`.
- Remove legacy local development certificate handling from `local.go`.
- Create a maintenance loop in `maintenance.go` for periodic DNS and certificate renewal tasks.
- Enhance DNS provider functionality in `provider.go` to sync A records with the public IP.
