+++ portal/acme/acme.go
@@ -2,49 +2,26 @@ package acme
 
 import (
 	"context"
-	"crypto"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
-	"crypto/x509"
-	"encoding/base64"
-	"encoding/pem"
 	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
-	"slices"
-	"strconv"
 	"strings"
 	"sync"
-	"time"
 
-	"github.com/go-acme/lego/v4/certcrypto"
 	"github.com/go-acme/lego/v4/certificate"
-	lego "github.com/go-acme/lego/v4/lego"
-	"github.com/go-acme/lego/v4/registration"
-	"github.com/rs/zerolog/log"
 
-	"github.com/gosuda/portal-tunnel/v2/portal/acme/internal/dnsrecord"
 	"github.com/gosuda/portal-tunnel/v2/portal/identity"
-	"github.com/gosuda/portal-tunnel/v2/portal/keyless"
 	"github.com/gosuda/portal-tunnel/v2/types"
 	"github.com/gosuda/portal-tunnel/v2/utils"
 )
 
 const (
-	fullChainFileName             = "fullchain.pem"
-	keyFileName                   = "privatekey.pem"
-	accountKeyFileName            = "acme-account.key"
-	registrationFileName          = "acme-registration.json"
-	ensGaslessHostnamesFileName   = "ens-gasless-hostnames.json"
-	gaslessENSTXTPrefix           = "ENS1 "
-	defaultENSGaslessResolver     = "0x238A8F792dFA6033814B18618aD4100654aeef01"
-	defaultACMEEmailPrefix        = "acme@"
-	defaultRenewInterval          = 24 * time.Hour
-	defaultManagedDNSSyncInterval = 3 * time.Hour
-	defaultDNSRetryInterval       = 10 * time.Minute
-	defaultSyncTimeout            = 2 * time.Minute
+	fullChainFileName      = "fullchain.pem"
+	keyFileName            = "privatekey.pem"
+	accountKeyFileName     = "acme-account.key"
+	registrationFileName   = "acme-registration.json"
+	defaultACMEEmailPrefix = "acme@"
 )
 
 type Config struct {
@@ -77,24 +54,6 @@ type Manager struct {
 	ensStatus   *utils.Snapshot[types.ENSStatus]
 	echCommands chan echDNSCommand
 	ensCommands chan ensDNSCommand
-}
-
-type echDNSCommand struct {
-	hostname string
-	record   dnsrecord.HTTPSRecord
-	remove   bool
-}
-
-type ensDNSCommand struct {
-	hostname string
-	address  string
-	remove   bool
-}
-
-type acmeUser struct {
-	Key          crypto.PrivateKey
-	Registration *registration.Resource
-	Email        string
 }
 
 func NewManager(cfg Config) (*Manager, error) {
@@ -160,61 +119,6 @@ func NewManager(cfg Config) (*Manager, error) {
 	return manager, nil
 }
 
-func newENSStatus(cfg Config, dns DNSProvider) types.ENSStatus {
-	provider := strings.TrimSpace(cfg.DNSProvider)
-	if dns != nil {
-		provider = dns.Name()
-	}
-	return types.ENSStatus{
-		Enabled:  cfg.ENSGaslessEnabled && !utils.IsLocalRelayHost(cfg.BaseDomain),
-		Provider: provider,
-		Address:  strings.TrimSpace(cfg.ENSGaslessAddress),
-	}
-}
-
-func (m *Manager) ENSStatus() types.ENSStatus {
-	if m == nil {
-		return types.ENSStatus{}
-	}
-	status := types.ENSStatus{}
-	if m.ensStatus != nil {
-		status = m.ensStatus.Load()
-	}
-	if status.Provider == "" && m.dns != nil {
-		status.Provider = m.dns.Name()
-	}
-	return status
-}
-
-func (m *Manager) setENSStatus(state, record, message string, syncErr error) {
-	if m == nil {
-		return
-	}
-	status := newENSStatus(m.cfg, m.dns)
-	status.DNSSECState = strings.TrimSpace(state)
-	status.DSRecord = strings.TrimSpace(record)
-	status.Message = strings.TrimSpace(message)
-	status.Verified = syncErr == nil && ensDNSSECVerified(status.DNSSECState)
-	if syncErr != nil {
-		status.LastError = syncErr.Error()
-	}
-
-	if m.ensStatus == nil {
-		m.ensStatus = utils.NewSnapshot(status)
-		return
-	}
-	m.ensStatus.Store(status)
-}
-
-func ensDNSSECVerified(state string) bool {
-	switch strings.ToLower(strings.TrimSpace(state)) {
-	case "active", "on", "signing", "transfer":
-		return true
-	default:
-		return false
-	}
-}
-
 func (m *Manager) EnsureCertificate(ctx context.Context) (string, string, error) {
 	if m == nil {
 		return "", "", errors.New("acme manager is nil")
@@ -239,7 +143,7 @@ func (m *Manager) EnsureCertificate(ctx context.Context) (string, string, error)
 	if manual {
 		return certFile, keyFile, nil
 	}
-	if !m.managedACME() {
+	if m.dns == nil {
 		return m.ensureManualCertificate()
 	}
 
@@ -279,7 +183,7 @@ func (m *Manager) EnsureTLSMaterial(ctx context.Context) ([]byte, []byte, error)
 }
 
 func (m *Manager) Start(ctx context.Context) {
-	if m == nil || utils.IsLocalRelayHost(m.cfg.BaseDomain) || (m.dns == nil && !m.cfg.ENSGaslessEnabled && !m.managedACME()) {
+	if m == nil || utils.IsLocalRelayHost(m.cfg.BaseDomain) || (m.dns == nil && !m.cfg.ENSGaslessEnabled) {
 		return
 	}
 
@@ -309,10 +213,6 @@ func (m *Manager) TLSFiles() (string, string, error) {
 		return "", "", errors.New("relay certificate files do not exist")
 	}
 	return certFile, keyFile, nil
-}
-
-func (m *Manager) managedACME() bool {
-	return m != nil && m.dns != nil
 }
 
 func (m *Manager) ensureManualCertificate() (string, string, error) {
@@ -397,552 +297,4 @@ func (m *Manager) provision(ctx context.Context) error {
 		return fmt.Errorf("write private key: %w", err)
 	}
 	return nil
-}
-
-func (m *Manager) maintenanceLoop(ctx context.Context) {
-	defer m.wg.Done()
-	pendingECH := make(map[string]echDNSCommand)
-	pendingENS := make(map[string]ensDNSCommand)
-	lastPublicIP := ""
-
-	renewTicker := time.NewTicker(defaultRenewInterval)
-	dnsTicker := time.NewTicker(defaultManagedDNSSyncInterval)
-	retryTicker := time.NewTicker(defaultDNSRetryInterval)
-	defer renewTicker.Stop()
-	defer dnsTicker.Stop()
-	defer retryTicker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.stopCh:
-			return
-		case command := <-m.echCommands:
-			syncCtx, cancel := context.WithTimeout(ctx, defaultSyncTimeout)
-			err := m.applyECHCommand(syncCtx, command)
-			cancel()
-			if err != nil {
-				pendingECH[command.hostname] = command
-				log.Warn().Err(err).Str("hostname", command.hostname).Msg("apply ECH dns command")
-			} else {
-				delete(pendingECH, command.hostname)
-			}
-		case command := <-m.ensCommands:
-			syncCtx, cancel := context.WithTimeout(ctx, defaultSyncTimeout)
-			err := m.applyENSCommand(syncCtx, command)
-			cancel()
-			if err != nil {
-				pendingENS[command.hostname] = command
-				log.Warn().Err(err).Str("hostname", command.hostname).Msg("apply ENS dns command")
-			} else {
-				delete(pendingENS, command.hostname)
-			}
-		case <-dnsTicker.C:
-			syncCtx, cancel := context.WithTimeout(ctx, defaultSyncTimeout)
-			err := m.syncDNS(syncCtx)
-			cancel()
-			if err != nil {
-				log.Warn().Err(err).Str("base_domain", m.cfg.BaseDomain).Msg("sync managed dns records")
-			}
-		case <-retryTicker.C:
-			syncCtx, cancel := context.WithTimeout(ctx, defaultSyncTimeout)
-			var err error
-			if m.cfg.ENSGaslessEnabled {
-				publicIP, resolveErr := utils.ResolvePublicIPv4(syncCtx)
-				if resolveErr != nil {
-					err = fmt.Errorf("detect public ip: %w", resolveErr)
-				} else if publicIP != lastPublicIP {
-					ensErr := m.syncTrackedENSGaslessHostARecords(syncCtx, publicIP)
-					err = errors.Join(err, ensErr)
-					if ensErr == nil {
-						lastPublicIP = publicIP
-					}
-				}
-				if !m.ENSStatus().Verified {
-					err = errors.Join(err, m.syncENSDNSSEC(syncCtx))
-				}
-			}
-		drainECH:
-			for {
-				select {
-				case command := <-m.echCommands:
-					pendingECH[command.hostname] = command
-				default:
-					break drainECH
-				}
-			}
-		drainENS:
-			for {
-				select {
-				case command := <-m.ensCommands:
-					pendingENS[command.hostname] = command
-				default:
-					break drainENS
-				}
-			}
-			for hostname, command := range pendingECH {
-				if commandErr := m.applyECHCommand(syncCtx, command); commandErr != nil {
-					err = errors.Join(err, commandErr)
-				} else {
-					delete(pendingECH, hostname)
-				}
-			}
-			for hostname, command := range pendingENS {
-				if commandErr := m.applyENSCommand(syncCtx, command); commandErr != nil {
-					err = errors.Join(err, commandErr)
-				} else {
-					delete(pendingENS, hostname)
-				}
-			}
-			cancel()
-			if err != nil {
-				log.Warn().Err(err).Str("base_domain", m.cfg.BaseDomain).Msg("sync dns records")
-			}
-		case <-renewTicker.C:
-			_, _, manual, err := m.manualCertificateOverride()
-			if err != nil || manual || !m.managedACME() || !m.shouldRenew() {
-				continue
-			}
-			renewCtx, cancel := context.WithTimeout(ctx, defaultSyncTimeout)
-			err = m.provision(renewCtx)
-			cancel()
-			if err != nil {
-				log.Warn().Err(err).Str("base_domain", m.cfg.BaseDomain).Msg("renew acme certificate")
-			}
-		}
-	}
-}
-
-func (m *Manager) syncDNS(ctx context.Context) error {
-	if m == nil || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
-		return nil
-	}
-	publicIP, err := utils.ResolvePublicIPv4(ctx)
-	if err != nil {
-		return fmt.Errorf("detect public ip: %w", err)
-	}
-	return m.syncDNSForIP(ctx, publicIP)
-}
-
-func (m *Manager) syncDNSForIP(ctx context.Context, publicIP string) error {
-	_, _, manual, err := m.manualCertificateOverride()
-	if err != nil {
-		return err
-	}
-	if manual || !m.managedACME() {
-		return nil
-	}
-	return m.dns.EnsureARecords(ctx, m.cfg.BaseDomain, publicIP)
-}
-
-func (m *Manager) SyncECHConfig(ctx context.Context, hostname string, echConfigList []byte, port int) error {
-	if m == nil || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
-		return nil
-	}
-	if m.dns == nil {
-		return nil
-	}
-	hostname = utils.NormalizeHostname(hostname)
-	if hostname == "" {
-		return errors.New("hostname is required")
-	}
-	if !utils.HostnameMatchesBaseDomain(hostname, m.cfg.BaseDomain) {
-		return fmt.Errorf("hostname %q is outside acme base domain %q", hostname, m.cfg.BaseDomain)
-	}
-	echConfigList, err := keyless.NormalizeEncryptedClientHelloConfigList(echConfigList)
-	if err != nil {
-		return err
-	}
-	if port < 0 || port > 65535 {
-		return errors.New("https record port must be between 0 and 65535")
-	}
-	svcParams := `ech="` + base64.StdEncoding.EncodeToString(echConfigList) + `"`
-	if port > 0 && port != 443 {
-		svcParams += " port=" + strconv.Itoa(port)
-	}
-	record := dnsrecord.HTTPSRecord{Priority: 1, Target: ".", SvcParams: svcParams}
-	record, err = record.Normalized()
-	if err != nil {
-		return err
-	}
-	return m.queueECHCommand(ctx, echDNSCommand{hostname: hostname, record: record})
-}
-
-func (m *Manager) DeleteECHConfig(ctx context.Context, hostname string) error {
-	if m == nil || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
-		return nil
-	}
-	if m.dns == nil {
-		return nil
-	}
-	hostname = utils.NormalizeHostname(hostname)
-	if hostname == "" {
-		return nil
-	}
-	if !utils.HostnameMatchesBaseDomain(hostname, m.cfg.BaseDomain) {
-		return nil
-	}
-
-	return m.queueECHCommand(ctx, echDNSCommand{hostname: hostname, remove: true})
-}
-
-func (m *Manager) queueECHCommand(ctx context.Context, command echDNSCommand) error {
-	select {
-	case m.echCommands <- command:
-		return nil
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-m.stopCh:
-		return errors.New("acme manager is stopped")
-	}
-}
-
-func (m *Manager) applyECHCommand(ctx context.Context, command echDNSCommand) error {
-	if command.remove {
-		if err := m.dns.DeleteHTTPSRecord(ctx, command.hostname); err != nil {
-			return err
-		}
-		if command.hostname != m.cfg.BaseDomain {
-			if err := m.dns.DeleteARecord(ctx, command.hostname); err != nil {
-				return fmt.Errorf("delete ECH A record for %s: %w", command.hostname, err)
-			}
-		}
-		return nil
-	}
-
-	publicIP, err := utils.ResolvePublicIPv4(ctx)
-	if err != nil {
-		return fmt.Errorf("detect public ip for ECH hostname %s: %w", command.hostname, err)
-	}
-	if err := m.dns.EnsureARecord(ctx, command.hostname, publicIP); err != nil {
-		return fmt.Errorf("ensure ECH A record for %s: %w", command.hostname, err)
-	}
-	return m.dns.EnsureHTTPSRecord(ctx, command.hostname, command.record)
-}
-
-func (m *Manager) syncENSGasless(ctx context.Context) error {
-	if m == nil || !m.cfg.ENSGaslessEnabled || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
-		return nil
-	}
-	if m.dns == nil {
-		return errors.New("ACME_DNS_PROVIDER is required")
-	}
-
-	if err := m.syncENSDNSSEC(ctx); err != nil {
-		return err
-	}
-	if err := m.SyncENSGaslessHostname(ctx, m.cfg.BaseDomain, m.cfg.ENSGaslessAddress); err != nil {
-		return fmt.Errorf("queue ens gasless txt: %w", err)
-	}
-	return nil
-}
-
-func (m *Manager) syncENSDNSSEC(ctx context.Context) error {
-	state, dsRecord, message, err := m.dns.EnsureDNSSEC(ctx, m.cfg.BaseDomain)
-	if err != nil {
-		m.setENSStatus("", "", "", err)
-		return fmt.Errorf("ensure dnssec: %w", err)
-	}
-	m.setENSStatus(state, dsRecord, message, nil)
-	return nil
-}
-
-func (m *Manager) SyncENSGaslessHostname(ctx context.Context, hostname, address string) error {
-	if m == nil || !m.cfg.ENSGaslessEnabled || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
-		return nil
-	}
-	if m.dns == nil {
-		return errors.New("ACME_DNS_PROVIDER is required")
-	}
-
-	hostname = utils.NormalizeHostname(hostname)
-	if hostname == "" {
-		return errors.New("hostname is required")
-	}
-	if !utils.HostnameMatchesBaseDomain(hostname, m.cfg.BaseDomain) {
-		return fmt.Errorf("hostname %q is outside acme base domain %q", hostname, m.cfg.BaseDomain)
-	}
-
-	address, err := identity.NormalizeEVMAddress(address)
-	if err != nil {
-		return fmt.Errorf("normalize ens gasless address: %w", err)
-	}
-	return m.queueENSCommand(ctx, ensDNSCommand{hostname: hostname, address: address})
-}
-
-func (m *Manager) DeleteENSGaslessHostname(ctx context.Context, hostname string) error {
-	if m == nil || !m.cfg.ENSGaslessEnabled || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
-		return nil
-	}
-	if m.dns == nil {
-		return errors.New("ACME_DNS_PROVIDER is required")
-	}
-
-	hostname = utils.NormalizeHostname(hostname)
-	if hostname == "" {
-		return nil
-	}
-	if !utils.HostnameMatchesBaseDomain(hostname, m.cfg.BaseDomain) {
-		return nil
-	}
-	if hostname == m.cfg.BaseDomain {
-		return nil
-	}
-	return m.queueENSCommand(ctx, ensDNSCommand{hostname: hostname, remove: true})
-}
-
-func (m *Manager) queueENSCommand(ctx context.Context, command ensDNSCommand) error {
-	select {
-	case m.ensCommands <- command:
-		return nil
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-m.stopCh:
-		return errors.New("acme manager is stopped")
-	}
-}
-
-func (m *Manager) applyENSCommand(ctx context.Context, command ensDNSCommand) error {
-	if command.remove {
-		if err := m.dns.DeleteTXTRecords(ctx, command.hostname, gaslessENSTXTPrefix); err != nil {
-			return err
-		}
-		if err := m.dns.DeleteARecord(ctx, command.hostname); err != nil {
-			return err
-		}
-		return m.updateTrackedENSGaslessHostnames(func(hostnames []string) []string {
-			return slices.DeleteFunc(hostnames, func(hostname string) bool { return hostname == command.hostname })
-		})
-	}
-
-	if command.hostname != m.cfg.BaseDomain {
-		publicIP, err := utils.ResolvePublicIPv4(ctx)
-		if err != nil {
-			return fmt.Errorf("detect public ip: %w", err)
-		}
-		if err := m.dns.EnsureARecord(ctx, command.hostname, publicIP); err != nil {
-			return fmt.Errorf("ensure ens gasless A record for %s: %w", command.hostname, err)
-		}
-	}
-	value := gaslessENSTXTPrefix + defaultENSGaslessResolver + " " + strings.TrimSpace(command.address)
-	if err := m.dns.EnsureTXTRecord(ctx, command.hostname, value); err != nil {
-		return err
-	}
-	return m.updateTrackedENSGaslessHostnames(func(hostnames []string) []string {
-		return append(hostnames, command.hostname)
-	})
-}
-
-func (m *Manager) reconcileTrackedENSGaslessHostnames(ctx context.Context) error {
-	if m == nil || !m.cfg.ENSGaslessEnabled || utils.IsLocalRelayHost(m.cfg.BaseDomain) || m.dns == nil {
-		return nil
-	}
-
-	var cleanupErr error
-	if err := m.updateTrackedENSGaslessHostnames(func(hostnames []string) []string {
-		remaining := hostnames[:0]
-		for _, hostname := range hostnames {
-			if err := m.dns.DeleteTXTRecords(ctx, hostname, gaslessENSTXTPrefix); err != nil {
-				remaining = append(remaining, hostname)
-				cleanupErr = errors.Join(cleanupErr, fmt.Errorf("delete ens gasless txt for %s: %w", hostname, err))
-				continue
-			}
-			if err := m.dns.DeleteARecord(ctx, hostname); err != nil {
-				remaining = append(remaining, hostname)
-				cleanupErr = errors.Join(cleanupErr, fmt.Errorf("delete ens gasless A record for %s: %w", hostname, err))
-			}
-		}
-		return remaining
-	}); err != nil {
-		cleanupErr = errors.Join(cleanupErr, fmt.Errorf("persist ens gasless hostnames: %w", err))
-	}
-	return cleanupErr
-}
-
-func (m *Manager) syncTrackedENSGaslessHostARecords(ctx context.Context, publicIP string) error {
-	hostnames, err := m.trackedENSGaslessHostnames()
-	if err != nil {
-		return err
-	}
-	var syncErr error
-	for _, hostname := range hostnames {
-		if err := m.dns.EnsureARecord(ctx, hostname, publicIP); err != nil {
-			syncErr = errors.Join(syncErr, fmt.Errorf("ensure ens gasless A record for %s: %w", hostname, err))
-		}
-	}
-	return syncErr
-}
-
-func (m *Manager) trackedENSGaslessHostnames() ([]string, error) {
-	if m == nil {
-		return nil, nil
-	}
-
-	path := filepath.Join(m.cfg.KeyDir, ensGaslessHostnamesFileName)
-	var hostnames []string
-	if _, err := utils.ReadJSONFileIfExists(path, &hostnames); err != nil {
-		return nil, err
-	}
-	return utils.NormalizeChildHostnames(hostnames, m.cfg.BaseDomain), nil
-}
-
-func (m *Manager) updateTrackedENSGaslessHostnames(update func([]string) []string) error {
-	if m == nil {
-		return nil
-	}
-
-	path := filepath.Join(m.cfg.KeyDir, ensGaslessHostnamesFileName)
-	hostnames, err := m.trackedENSGaslessHostnames()
-	if err != nil {
-		return err
-	}
-	if update != nil {
-		hostnames = utils.NormalizeChildHostnames(update(hostnames), m.cfg.BaseDomain)
-	}
-	if len(hostnames) == 0 {
-		if err := os.Remove(path); err != nil && !errors.Is(err, os.ErrNotExist) {
-			return err
-		}
-		return nil
-	}
-	return utils.WriteJSONFile(path, hostnames, 0o600)
-}
-
-func (m *Manager) shouldRenew() bool {
-	certFile := filepath.Join(m.cfg.KeyDir, fullChainFileName)
-	needsRenewal, err := certNeedsRenewal(certFile, certificateDomains(m.cfg.BaseDomain))
-	return err == nil && needsRenewal
-}
-
-func certificateDomains(baseDomain string) []string {
-	return []string{baseDomain, "*." + baseDomain}
-}
-
-func certNeedsRenewal(certFile string, domains []string) (bool, error) {
-	cert, err := loadCertificate(certFile)
-	if err != nil {
-		return false, err
-	}
-	if time.Until(cert.NotAfter) < 30*24*time.Hour {
-		return true, nil
-	}
-	return !certificateCoversDomains(cert, domains), nil
-}
-
-func certCoversDomains(certFile string, domains []string) (bool, error) {
-	cert, err := loadCertificate(certFile)
-	if err != nil {
-		return false, err
-	}
-	return certificateCoversDomains(cert, domains), nil
-}
-
-func loadCertificate(certFile string) (*x509.Certificate, error) {
-	certPEM, err := os.ReadFile(certFile)
-	if err != nil {
-		return nil, err
-	}
-	return utils.ParseCertificatePEM(certPEM)
-}
-
-func certificateCoversDomains(cert *x509.Certificate, domains []string) bool {
-	for _, domain := range domains {
-		if wildcardDomain, ok := strings.CutPrefix(domain, "*."); ok {
-			if !certificateCoversHostname(cert, "probe."+wildcardDomain) {
-				return false
-			}
-			continue
-		}
-		if !certificateCoversHostname(cert, domain) {
-			return false
-		}
-	}
-	return true
-}
-
-func certificateCoversHostname(cert *x509.Certificate, hostname string) bool {
-	return cert != nil && cert.VerifyHostname(hostname) == nil
-}
-
-func newClient(ctx context.Context, email, accountKeyFile, registrationFile string, dnsProvider DNSProvider) (*lego.Client, error) {
-	accountKey, err := loadOrCreateAccountKey(accountKeyFile)
-	if err != nil {
-		return nil, fmt.Errorf("load acme account key: %w", err)
-	}
-
-	var accountReg registration.Resource
-	accountRegPtr := (*registration.Resource)(nil)
-	if ok, err := utils.ReadJSONFileIfExists(registrationFile, &accountReg); err != nil {
-		return nil, fmt.Errorf("load acme registration: %w", err)
-	} else if ok {
-		accountRegPtr = &accountReg
-	}
-
-	user := &acmeUser{
-		Email:        email,
-		Key:          accountKey,
-		Registration: accountRegPtr,
-	}
-
-	clientConfig := lego.NewConfig(user)
-	clientConfig.CADirURL = lego.LEDirectoryProduction
-	clientConfig.Certificate.KeyType = certcrypto.RSA2048
-
-	client, err := lego.NewClient(clientConfig)
-	if err != nil {
-		return nil, fmt.Errorf("create acme client: %w", err)
-	}
-
-	if dnsProvider == nil {
-		return nil, errors.New("ACME_DNS_PROVIDER is required")
-	}
-	challengeProvider, err := dnsProvider.ChallengeProvider(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("create dns challenge provider: %w", err)
-	}
-	if err := client.Challenge.SetDNS01Provider(challengeProvider); err != nil {
-		return nil, fmt.Errorf("set dns01 provider: %w", err)
-	}
-
-	if user.Registration == nil {
-		reg, err := client.Registration.Register(registration.RegisterOptions{TermsOfServiceAgreed: true})
-		if err != nil {
-			return nil, fmt.Errorf("register acme account: %w", err)
-		}
-		user.Registration = reg
-		if err := utils.WriteJSONFile(registrationFile, reg, 0o600); err != nil {
-			return nil, fmt.Errorf("persist acme registration: %w", err)
-		}
-	}
-
-	return client, nil
-}
-
-func (u *acmeUser) GetEmail() string                        { return u.Email }
-func (u *acmeUser) GetRegistration() *registration.Resource { return u.Registration }
-func (u *acmeUser) GetPrivateKey() crypto.PrivateKey        { return u.Key }
-
-func loadOrCreateAccountKey(path string) (crypto.PrivateKey, error) {
-	keyPEM, err := os.ReadFile(path)
-	if err == nil {
-		return utils.ParsePrivateKeyPEM(keyPEM)
-	}
-	if !errors.Is(err, os.ErrNotExist) {
-		return nil, err
-	}
-
-	key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	if err != nil {
-		return nil, fmt.Errorf("generate account key: %w", err)
-	}
-	pkcs8, err := x509.MarshalPKCS8PrivateKey(key)
-	if err != nil {
-		return nil, fmt.Errorf("marshal account key: %w", err)
-	}
-	keyPEM = pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: pkcs8})
-	if err := utils.WriteFileAtomic(path, keyPEM, 0o600); err != nil {
-		return nil, fmt.Errorf("persist account key: %w", err)
-	}
-	return key, nil
 }
