--- portal/acme/local.go
@@ -1,105 +1,0 @@
-package acme
-
-import (
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/pem"
-	"fmt"
-	"math/big"
-	"net"
-	"path/filepath"
-	"time"
-
-	"github.com/gosuda/portal-tunnel/v2/utils"
-)
-
-const localDevelopmentCertificateTTL = 3650 * 24 * time.Hour
-
-func ensureLocalDevelopmentCertificate(keyDir, baseHost string) error {
-	domains := localDevelopmentDomains(baseHost)
-	keyFile := filepath.Join(keyDir, keyFileName)
-	certFile := filepath.Join(keyDir, fullChainFileName)
-
-	if utils.FileExists(keyFile) && utils.FileExists(certFile) {
-		covered, err := certCoversDomains(certFile, domains)
-		if err == nil && covered {
-			return nil
-		}
-	}
-
-	if err := utils.EnsureParentDir(keyFile); err != nil {
-		return err
-	}
-	if err := utils.EnsureParentDir(certFile); err != nil {
-		return err
-	}
-
-	privateKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	if err != nil {
-		return fmt.Errorf("generate local dev private key: %w", err)
-	}
-
-	serialLimit := new(big.Int).Lsh(big.NewInt(1), 128)
-	serialNumber, err := rand.Int(rand.Reader, serialLimit)
-	if err != nil {
-		return fmt.Errorf("generate local dev certificate serial: %w", err)
-	}
-
-	now := time.Now().UTC()
-	template := &x509.Certificate{
-		SerialNumber: serialNumber,
-		Subject: pkix.Name{
-			CommonName:   baseHost,
-			Organization: []string{"Portal Local Development"},
-		},
-		NotBefore:             now.Add(-1 * time.Hour),
-		NotAfter:              now.Add(localDevelopmentCertificateTTL),
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment | x509.KeyUsageKeyAgreement | x509.KeyUsageCertSign,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
-		BasicConstraintsValid: true,
-		IsCA:                  true,
-	}
-
-	for _, domain := range domains {
-		if ip := net.ParseIP(domain); ip != nil {
-			template.IPAddresses = append(template.IPAddresses, ip)
-			continue
-		}
-		template.DNSNames = append(template.DNSNames, domain)
-	}
-
-	certDER, err := x509.CreateCertificate(rand.Reader, template, template, &privateKey.PublicKey, privateKey)
-	if err != nil {
-		return fmt.Errorf("create local dev certificate: %w", err)
-	}
-
-	certPEM := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: certDER})
-	keyDER, err := x509.MarshalPKCS8PrivateKey(privateKey)
-	if err != nil {
-		return fmt.Errorf("marshal local dev private key: %w", err)
-	}
-	keyPEM := pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: keyDER})
-
-	if err := utils.WriteFileAtomic(certFile, certPEM, 0o644); err != nil {
-		return fmt.Errorf("write local dev certificate: %w", err)
-	}
-	if err := utils.WriteFileAtomic(keyFile, keyPEM, 0o600); err != nil {
-		return fmt.Errorf("write local dev private key: %w", err)
-	}
-	return nil
-}
-
-func localDevelopmentDomains(baseHost string) []string {
-	baseHost = utils.NormalizeBaseDomain(baseHost)
-	domains := []string{"localhost", "*.localhost", "127.0.0.1", "::1"}
-	if baseHost != "" && baseHost != "localhost" {
-		domains = append(domains, baseHost)
-		if net.ParseIP(baseHost) == nil {
-			domains = append(domains, "*."+baseHost)
-		}
-	}
-	return domains
-}
