+++ portal/acme/provider.go
@@ -14,6 +14,7 @@ import (
 	"github.com/gosuda/portal-tunnel/v2/portal/acme/njalla"
 	"github.com/gosuda/portal-tunnel/v2/portal/acme/route53"
 	"github.com/gosuda/portal-tunnel/v2/portal/acme/vultr"
+	"github.com/gosuda/portal-tunnel/v2/utils"
 )
 
 const (
@@ -67,4 +68,22 @@ func NewDNSProvider(providerType string, cfg Config) (DNSProvider, error) {
 	default:
 		return nil, fmt.Errorf("unsupported acme dns provider: %q", providerType)
 	}
+}
+
+func (m *Manager) syncDNS(ctx context.Context) error {
+	if m == nil || utils.IsLocalRelayHost(m.cfg.BaseDomain) {
+		return nil
+	}
+	publicIP, err := utils.ResolvePublicIPv4(ctx)
+	if err != nil {
+		return fmt.Errorf("detect public ip: %w", err)
+	}
+	_, _, manual, err := m.manualCertificateOverride()
+	if err != nil {
+		return err
+	}
+	if manual || m.dns == nil {
+		return nil
+	}
+	return m.dns.EnsureARecords(ctx, m.cfg.BaseDomain, publicIP)
 }
